feat: Implementar ORM con SQLAlchemy y optimizar gestión de ingredientes

+++ IMenu.py
@@ -2,12 +2,13 @@
 from typing import Protocol, List, Optional
 from Ingrediente import Ingrediente
 from Stock import Stock
+from decimal import Decimal
 
 class IMenu(Protocol):
     """Interfaz para los elementos del menú."""
     nombre: str
     ingredientes: List[Ingrediente]
-    precio: float
+    precio: Decimal
     cantidad: int
     icono_path: Optional[str]
